# backend/src/openapi_server/models/object.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class Object(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Object - a model defined in OpenAPIf

        id: The id of this Object.
        description: The description of this Object.
        value: The value of this Object.
    """

    id: str = Field(alias="id")
    description: str = Field(alias="description")
    value: str = Field(alias="value")

    @validator("id")
    def id_pattern(cls, value):
        assert value is not None and re.match(r"^.{0,20}", value)
        return value

    @validator("description")
    def description_pattern(cls, value):
        assert value is not None and re.match(r"^.{0,40}", value)
        return value

    @validator("value")
    def value_pattern(cls, value):
        assert value is not None and re.match(r"^.{0,20}", value)
        return value

Object.update_forward_refs()